refactor(auth-form): extract all logic into useAuthForm — component is now a thin shell
Move every state/effect/derived/handler (the full 105-572 block) plus
AuthMode/MODE_HEADING_* /AuthFormProps verbatim into use-auth-form.ts;
AuthForm is now `const {...} = useAuthForm(props); return (<render>)`.
The ~470-line render JSX is byte-identical (sed-copied, zero diff) so
behavior is preserved by construction — this is a pure relocation.

Verified: tsc clean (render fully resolves against the hook return),
12 RTL characterization tests pass unchanged, full suite 93/93, build OK.
auth-form.tsx 1046 -> 623 LOC (thin shell); logic in use-auth-form.ts.

Residual: the net covers signin/signup/forgot/passkey/empty-submit/
MFA-gate/error but NOT live MFA-code verify, resend-verification, or
OAuth-callback params. Those moved verbatim (behavior preserved by
construction) but have no automated test — smoke-test before merge.
